requirements.txt
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
arch==6.2.0
statsmodels==0.14.0
vectorbt==0.25.4
yfinance==0.2.28
ccxt==4.0.80
alpaca-trade-api==3.0.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
PyPortfolioOpt==1.5.5
hmmlearn==0.3.0
pykalman==0.9.5
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
websockets<11,>=9.0
jupyter==1.0.0
